+++ src/fhir/sdc-import-common.js
@@ -846,6 +846,23 @@ function addCommonSDCImportFns(ns) {
   };
 
 
+  /**
+   * Parse questionnaire item for "hidden" extension
+   *
+   * @param lfItem {object} - LForms item object to be assigned the isHiddenInDef flag if the item is to be hidden.
+   * @param qItem {object} - Questionnaire item object
+   * @private
+   * @return true if the item is hidden or if its ancestor is hidden, false otherwise
+   */
+  self._processHiddenItem = function(lfItem, qItem) {
+    var ci = LForms.Util.findObjectInArray(qItem.extension, 'url', self.fhirExtUrlHidden);
+    if(ci) {
+      lfItem.isHiddenInDef = typeof ci.valueBoolean === 'boolean'? ci.valueBoolean: ci.valueBoolean === 'true';
+    }
+    return lfItem.isHiddenInDef;
+  };
+
+
   // ---------------- QuestionnaireResponse Import ---------------
 
   var qrImport = self._mergeQR;